+++ ext_tables.sql
@@ -1,0 +1,4 @@
+CREATE TABLE sys_domain (
+    domainName                            VARCHAR(255) NOT NULL DEFAULT '',
+    tx_bartacusplatformsh_routeDomainName VARCHAR(255) NOT NULL DEFAULT ''
+);